feat(dependencies): add new dependencies for image processing and OCR, update package versions for improved functionality and compatibility

+++ apps/admin/src/lib/actions/user.ts
@@ -1,5 +1,5 @@
 "use server";
-import { createSessionClient, createAdminClient, } from "@repo/api/server";
+import { createSessionClient, createAdminClient } from "@repo/api/server";
 import { create } from "domain";
 import { headers, cookies } from "next/headers";
 import { redirect } from "next/navigation";
@@ -62,7 +62,7 @@ async function getCurrentSession() {
 
 export async function getUserById(userId: string): Promise<Users | null> {
     try {
-        const { db } = await createAdminClient();
+        const { db } = await createSessionClient();
         const user = await db.getRow<Users>('app', 'user', userId);
         return user;
     } catch (error) {
@@ -73,7 +73,7 @@ export async function getUserById(userId: string): Promise<Users | null> {
 
 async function signIn(email: string) {
     try {
-        const { account } = await createAdminClient();
+        const { account } = await createSessionClient();
         const user = await account.createMagicURLToken(ID.unique(), email, `${BASE_URL}/auth/callback`);
         return user;
     } catch (error) {
@@ -83,7 +83,7 @@ async function signIn(email: string) {
 }
 
 async function signInWithOauth() {
-	const { account } = await createAdminClient();
+	const { account } = await createSessionClient();
 
   const origin = (await headers()).get("origin");
   
@@ -232,9 +232,21 @@ export async function signOut(): Promise<void> {
   }
 
 export async function deleteUserData() {
-    const { account } = await createSessionClient();
+    const { account, teams } = await createSessionClient();
     const { users, db } = await createAdminClient();
     const user = await account.get();
+    if (!user) {
+        return false;
+    }
+    
+    const userTeams = await teams.list();
+
+    //Must be admin to delete user
+    if (!userTeams.teams.some((team: any) => team.name === 'Admin')) {
+        return false;
+    }
+
+
     const deletedUserDoc = await db.deleteRow('app', 'user', user.$id);
     if (deletedUserDoc) {
         const deletedUser = await users.delete(user.$id);
